AutoBone: Add new dedicated AutoBone window
