+++ src/main/java/io/eiren/gui/SkeletonConfig.java
@@ -3,46 +3,33 @@ package io.eiren.gui;
 import java.awt.event.ItemEvent;
 import java.awt.event.ItemListener;
 import java.awt.event.MouseEvent;
-import java.io.File;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Map.Entry;
-import java.util.concurrent.Future;
 
 import javax.swing.JButton;
 import javax.swing.JCheckBox;
 import javax.swing.JLabel;
 import javax.swing.event.MouseInputAdapter;
 
-import io.eiren.gui.autobone.PoseRecorder;
-import org.apache.commons.lang3.tuple.Pair;
-
-import io.eiren.gui.autobone.AutoBone;
-import io.eiren.gui.autobone.PoseFrame;
-import io.eiren.gui.autobone.PoseRecordIO;
+import io.eiren.gui.autobone.AutoBoneWindow;
 import io.eiren.util.StringUtils;
 import io.eiren.util.ann.ThreadSafe;
-import io.eiren.util.collections.FastList;
 import io.eiren.vr.VRServer;
 import io.eiren.vr.processor.HumanSkeletonWithLegs;
 import io.eiren.vr.processor.HumanSkeleton;
-import io.eiren.util.logging.LogManager;
 
 public class SkeletonConfig extends EJBag {
 
 	private final VRServer server;
 	private final VRServerGUI gui;
-	private final PoseRecorder poseRecorder;
-	private final AutoBone autoBone;
-	private Thread autoBoneThread = null;
+	private final AutoBoneWindow autoBone;
 	private Map<String, SkeletonLabel> labels = new HashMap<>();
 
 	public SkeletonConfig(VRServer server, VRServerGUI gui) {
 		super();
 		this.server = server;
 		this.gui = gui;
-		this.poseRecorder = new PoseRecorder(server);
-		this.autoBone = new AutoBone(server);
+		this.autoBone = new AutoBoneWindow(server, this);
 
 		setAlignmentY(TOP_ALIGNMENT);
 		server.humanPoseProcessor.addSkeletonUpdatedCallback(this::skeletonUpdated);
@@ -111,169 +98,8 @@ public class SkeletonConfig extends EJBag {
 				addMouseListener(new MouseInputAdapter() {
 					@Override
 					public void mouseClicked(MouseEvent e) {
-						// Prevent running multiple times
-						if (autoBoneThread != null) {
-							return;
-						}
-
-						Thread thread = new Thread() {
-							@Override
-							public void run() {
-								try {
-									FastList<Pair<String, PoseFrame[]>> frameRecordings = new FastList<Pair<String, PoseFrame[]>>();
-
-									File loadFolder = new File("LoadRecordings");
-									if (loadFolder.isDirectory()) {
-										setText("Load");
-
-										File[] files = loadFolder.listFiles();
-										if (files != null) {
-											for (File file : files) {
-												if (file.isFile() && org.apache.commons.lang3.StringUtils.endsWithIgnoreCase(file.getName(), ".abf")) {
-													LogManager.log.info("[AutoBone] Detected recording at \"" + file.getPath() + "\", loading frames...");
-													PoseFrame[] frames = PoseRecordIO.readFromFile(file);
-
-													if (frames == null) {
-														LogManager.log.severe("Reading frames from \"" + file.getPath() + "\" failed...");
-													} else {
-														frameRecordings.add(Pair.of(file.getName(), frames));
-													}
-												}
-											}
-										}
-									}
-
-									if (frameRecordings.size() > 0) {
-										setText("Wait");
-										LogManager.log.info("[AutoBone] Done loading frames!");
-									} else if (poseRecorder.isReadyToRecord()) {
-										setText("Move");
-										// 1000 samples at 20 ms per sample is 20 seconds
-										int sampleCount = server.config.getInt("autobone.sampleCount", 1000);
-										long sampleRate = server.config.getLong("autobone.sampleRateMs", 20L);
-										Future<PoseFrame[]> framesFuture = poseRecorder.startFrameRecording(sampleCount, sampleRate);
-										PoseFrame[] frames = framesFuture.get();
-										LogManager.log.info("[AutoBone] Done recording!");
-
-										setText("Wait");
-										if (server.config.getBoolean("autobone.saveRecordings", true)) {
-											File saveFolder = new File("Recordings");
-											if (saveFolder.isDirectory() || saveFolder.mkdirs()) {
-												File saveRecording;
-												int recordingIndex = 1;
-												do {
-													saveRecording = new File(saveFolder, "ABRecording" + recordingIndex++ + ".abf");
-												} while (saveRecording.exists());
-
-												LogManager.log.info("[AutoBone] Exporting frames to \"" + saveRecording.getPath() + "\"...");
-												if (PoseRecordIO.writeToFile(saveRecording, frames)) {
-													LogManager.log.info("[AutoBone] Done exporting! Recording can be found at \"" + saveRecording.getPath() + "\".");
-												} else {
-													LogManager.log.severe("[AutoBone] Failed to export the recording to \"" + saveRecording.getPath() + "\".");
-												}
-											} else {
-												LogManager.log.severe("[AutoBone] Failed to create the recording directory \"" + saveFolder.getPath() + "\".");
-											}
-										}
-										frameRecordings.add(Pair.of("<Recording>", frames));
-									} else {
-										LogManager.log.severe("[AutoBone] No recordings found in \"" + loadFolder.getPath() + "\" and unable to record...");
-										return;
-									}
-
-									LogManager.log.info("[AutoBone] Processing frames...");
-									FastList<Float> heightPercentError = new FastList<Float>(frameRecordings.size());
-									for (Pair<String, PoseFrame[]> recording : frameRecordings) {
-										LogManager.log.info("[AutoBone] Processing frames from \"" + recording.getKey() + "\"...");
-										autoBone.reloadConfigValues();
-
-										autoBone.minDataDistance = server.config.getInt("autobone.minimumDataDistance", autoBone.minDataDistance);
-										autoBone.maxDataDistance = server.config.getInt("autobone.maximumDataDistance", autoBone.maxDataDistance);
-										autoBone.numEpochs = server.config.getInt("autobone.epochCount", autoBone.numEpochs);
-										autoBone.initialAdjustRate = server.config.getFloat("autobone.adjustRate", autoBone.initialAdjustRate);
-										autoBone.adjustRateDecay = server.config.getFloat("autobone.adjustRateDecay", autoBone.adjustRateDecay);
-										autoBone.slideErrorFactor = server.config.getFloat("autobone.slideErrorFactor", autoBone.slideErrorFactor);
-										autoBone.offsetErrorFactor = server.config.getFloat("autobone.offsetErrorFactor", autoBone.offsetErrorFactor);
-										autoBone.proportionErrorFactor = server.config.getFloat("autobone.proportionErrorFactor", autoBone.proportionErrorFactor);
-										autoBone.heightErrorFactor = server.config.getFloat("autobone.heightErrorFactor", autoBone.heightErrorFactor);
-
-										boolean calcInitError = server.config.getBoolean("autobone.calculateInitialError", true);
-										float targetHeight = server.config.getFloat("autobone.manualTargetHeight", -1f);
-										heightPercentError.add(autoBone.processFrames(recording.getValue(), calcInitError, targetHeight));
-
-										LogManager.log.info("[AutoBone] Done processing!");
-
-										//#region Stats/Values
-										Float neckLength = autoBone.getConfig("Neck");
-										Float chestLength = autoBone.getConfig("Chest");
-										Float waistLength = autoBone.getConfig("Waist");
-										Float hipWidth = autoBone.getConfig("Hips width");
-										Float legsLength = autoBone.getConfig("Legs length");
-										Float kneeHeight = autoBone.getConfig("Knee height");
-
-										float neckWaist = neckLength != null && waistLength != null ? neckLength / waistLength : 0f;
-										float chestWaist = chestLength != null && waistLength != null ? chestLength / waistLength : 0f;
-										float hipWaist = hipWidth != null && waistLength != null ? hipWidth / waistLength : 0f;
-										float legWaist = legsLength != null && waistLength != null ? legsLength / waistLength : 0f;
-										float legBody = legsLength != null && waistLength != null && neckLength != null ? legsLength / (waistLength + neckLength) : 0f;
-										float kneeLeg = kneeHeight != null && legsLength != null ? kneeHeight / legsLength : 0f;
-
-										LogManager.log.info("[AutoBone] Ratios: [{Neck-Waist: " + StringUtils.prettyNumber(neckWaist) +
-										"}, {Chest-Waist: " + StringUtils.prettyNumber(chestWaist) +
-										"}, {Hip-Waist: " + StringUtils.prettyNumber(hipWaist) +
-										"}, {Leg-Waist: " + StringUtils.prettyNumber(legWaist) +
-										"}, {Leg-Body: " + StringUtils.prettyNumber(legBody) +
-										"}, {Knee-Leg: " + StringUtils.prettyNumber(kneeLeg) + "}]");
-
-										boolean first = true;
-										StringBuilder configInfo = new StringBuilder("[");
-
-										for (Entry<String, Float> entry : autoBone.configs.entrySet()) {
-											if (!first) {
-												configInfo.append(", ");
-											} else {
-												first = false;
-											}
-
-											configInfo.append("{" + entry.getKey() + ": " + StringUtils.prettyNumber(entry.getValue()) + "}");
-										}
-
-										configInfo.append(']');
-
-										LogManager.log.info("[AutoBone] Length values: " + configInfo.toString());
-									}
-
-									if (heightPercentError.size() > 0) {
-										float mean = 0f;
-										for (float val : heightPercentError) {
-											mean += val;
-										}
-										mean /= heightPercentError.size();
-
-										float std = 0f;
-										for (float val : heightPercentError) {
-											float stdVal = val - mean;
-											std += stdVal * stdVal;
-										}
-										std = (float)Math.sqrt(std / heightPercentError.size());
-
-										LogManager.log.info("[AutoBone] Average height error: " + StringUtils.prettyNumber(mean, 6) + " (SD " + StringUtils.prettyNumber(std, 6) + ")");
-									}
-									//#endregion
-
-									// Update GUI values after adjustment
-									refreshAll();
-								} catch (Exception e1) {
-									LogManager.log.severe("[AutoBone] Failed adjustment!", e1);
-								} finally {
-									setText("Auto");
-									autoBoneThread = null;
-								}
-							}
-						};
-
-						autoBoneThread = thread;
-						thread.start();
+						autoBone.setVisible(true);
+						autoBone.toFront();
 					}
 				});
 			}}, s(c(4, row, 1), 3, 1));
